Support L2 wbinval
A27 and AX27 have L2 cache, but lack of coherent agant. It will cause
MAC driver dhcp fail, so it is necessary to add wbinval cache_ops
in cache uclass driver.

Also add a coherent flag which will parse the dma-coherent in mac node
of device tree to distinguish if it is necessary to flush and invalidate
cache during network data transactions. Meanwhile if dma-coherent is exist,
don't use PMA to configure MAC'desc buffer.