riscv: ae350: Move the DTB in front of kernel (#50)
Originally, u-boot-spl will place DTB directly after the kernel,
but the size of the kernel do not include the BSS section, which
means u-boot-spl will place the DTB in the kernel BSS section
causing the DTB to be cleared by kernel BSS initializtion.

Moving DTB in front of the kernel can avoid the error.

Signed-off-by: Dylan Jhong <dylan@andestech.com>
Reviewed-by: Rick Chen <rick@andestech.com>
Reviewed-by: Leo Yu-Chi Liang <ycliang@andestech.com>
Reviewed-on: http://gitea.andestech.com:3000/RD-SW/u-boot.official/pulls/50