riscv: Rename Andes cpu and board names (#60)
We try to tackle the non-coherent bitmap setting by
splitting the configs into 25-series and 45-series.
Use 25-series defconfigis for non-coherent bitmaps and
45-series for the rest, but this is not a good practice
as it introduces complexity but does not account for
coherent 25 SMP bitmaps.

So unify the 25/45 defconfigs and rename Andes cpu to
"andesv5" and board name to "ae350".

As for non-coherent bitmap, we will support non-coherent
UP bitmap by disabling cache and we might provide patches
for non-coherent SMP bitmap if requested.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/u-boot.official/pulls/60